WIP on moving all graphics API calls behind a "GAPI" interface. So far, the render loop, textures, and 3D model data have been migrated behind the interface. Still need to deal with Skyboxes and Shaders though.
Moving graphics commands behind an interface like this makes it easier to support other graphics APIs in the future (D3D, Vulkan, etc).

+++ Source/Rendering/Material.cpp
@@ -61,7 +61,7 @@ void Material::Activate(const Matrix4& objectToWorldMatrix)
     }
     
     // Set user-defined textures.
-    int textureUnit = 0;
+    uint8_t textureUnit = 0;
     for(auto& entry : mTextures)
     {
         if(entry.second != nullptr)
